move addmountains method to movingBg.js
