+++ src/config/gameOptions.js
